# server/api-server/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SpacetimeWiki REST API",
    "description": "Programmatic CRUD and search for pages, collections, and more. SpacetimeWiki is a real-time wiki with rich text editing, hierarchical collections, tagging, search, SSO/OAuth/LDAP auth, SCIM provisioning, WebAuthn/passkeys, import/export, and more.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/webauthn/auth/begin": {
      "get": {
        "tags": [
          "webauthn"
        ],
        "summary": "Auth Begin",
        "description": "Generate WebAuthn authentication options.\n\nReturns PublicKeyCredentialRequestOptions as JSON that the browser\nshould pass to navigator.credentials.get().\n\nIf email is provided, only allow credentials for that user.",
        "operationId": "auth_begin_api_v1_webauthn_auth_begin_get",
        "parameters": [
          {
            "name": "email",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Email"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WebAuthnBeginAuthResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/webauthn/auth/complete": {
      "post": {
        "tags": [
          "webauthn"
        ],
        "summary": "Auth Complete",
        "description": "Verify a WebAuthn authentication assertion.\n\nExpects the CredentialAssertionResponse from the browser:\n{\n    id: string,\n    rawId: string (base64url),\n    type: \"public-key\",\n    response: {\n        clientDataJSON: string (base64url),\n        authenticatorData: string (base64url),\n        signature: string (base64url),\n        userHandle: string (base64url, optional),\n    },\n}",
        "operationId": "auth_complete_api_v1_webauthn_auth_complete_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": true,
                "type": "object",
                "title": "Body"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WebAuthnAuthCompleteResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/webauthn/register/begin": {
      "get": {
        "tags": [
          "webauthn"
        ],
        "summary": "Register Begin",
        "description": "Generate WebAuthn registration options for a new credential.\n\nReturns PublicKeyCredentialCreationOptions as JSON that the browser\nshould pass to navigator.credentials.create().",
        "operationId": "register_begin_api_v1_webauthn_register_begin_get",
        "parameters": [
          {
            "name": "email",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Email"
            }
          },
          {
            "name": "display_name",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Display Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WebAuthnBeginRegisterResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/webauthn/register/complete": {
      "post": {
        "tags": [
          "webauthn"
        ],
        "summary": "Register Complete",
        "description": "Verify and store a WebAuthn registration credential.\n\nExpects the CredentialCreationResponse from the browser:\n{\n    id: string,\n    rawId: string (base64url),\n    type: \"public-key\",\n    response: {\n        clientDataJSON: string (base64url),\n        attestationObject: string (base64url),\n        transports: string[],\n    },\n    user_id: string,   // wiki user ID (passed through by the client)\n    email: string,      // for challenge lookup\n}",
        "operationId": "register_complete_api_v1_webauthn_register_complete_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": true,
                "type": "object",
                "title": "Body"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WebAuthnRegisterCompleteResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "WebAuthnAuthCompleteResponse": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status",
            "description": "Authentication status",
            "default": "ok"
          },
          "user": {
            "additionalProperties": true,
            "type": "object",
            "title": "User",
            "description": "Authenticated user data"
          }
        },
        "type": "object",
        "required": [
          "user"
        ],
        "title": "WebAuthnAuthCompleteResponse"
      },
      "WebAuthnBeginAuthResponse": {
        "properties": {
          "challenge": {
            "type": "string",
            "title": "Challenge",
            "description": "Authentication challenge"
          },
          "timeout": {
            "type": "integer",
            "title": "Timeout",
            "description": "Timeout in ms",
            "default": 300000
          },
          "rpId": {
            "type": "string",
            "title": "Rpid",
            "description": "Relying party ID",
            "default": "localhost"
          },
          "allowCredentials": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Allowcredentials"
          },
          "userVerification": {
            "type": "string",
            "title": "Userverification",
            "default": "preferred"
          }
        },
        "type": "object",
        "required": [
          "challenge"
        ],
        "title": "WebAuthnBeginAuthResponse"
      },
      "WebAuthnBeginRegisterResponse": {
        "properties": {
          "challenge": {
            "type": "string",
            "title": "Challenge",
            "description": "Registration challenge"
          },
          "rp": {
            "additionalProperties": true,
            "type": "object",
            "title": "Rp",
            "description": "Relying party info"
          },
          "user": {
            "additionalProperties": true,
            "type": "object",
            "title": "User",
            "description": "User info"
          },
          "pubKeyCredParams": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Pubkeycredparams",
            "description": "Accepted key types"
          },
          "timeout": {
            "type": "integer",
            "title": "Timeout",
            "description": "Timeout in ms",
            "default": 300000
          },
          "attestation": {
            "type": "string",
            "title": "Attestation",
            "description": "Attestation preference",
            "default": "none"
          },
          "excludeCredentials": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Excludecredentials"
          },
          "authenticatorSelection": {
            "additionalProperties": true,
            "type": "object",
            "title": "Authenticatorselection"
          }
        },
        "type": "object",
        "required": [
          "challenge",
          "rp",
          "user",
          "pubKeyCredParams"
        ],
        "title": "WebAuthnBeginRegisterResponse"
      },
      "WebAuthnRegisterCompleteResponse": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status",
            "description": "Registration status",
            "default": "ok"
          },
          "credential_id": {
            "type": "string",
            "title": "Credential Id",
            "description": "Registered credential ID",
            "default": ""
          }
        },
        "type": "object",
        "title": "WebAuthnRegisterCompleteResponse"
      }
    }
  },
  "servers": [
    {
      "url": "/",
      "description": "SpacetimeWiki API server"
    }
  ]
}